Add diagram of the revised version of The Strix
A few changes were made to the project design. Our AI model will be running on a local computer which is typically where CCTV feed is stored on. The doppler sensors will trigger the AI model and specify the camera which is supposed to be analyzed and then, the AI model will use image processing on thet specific camera feed and look for human presence, and if any humans are present, an alarm sequence will initiate. This will be an efficient implementation since the model will not be constantly analyzing the feeds.